Move all daemons to __init__
