+++ setup.cfg
@@ -9,11 +9,11 @@ packages =
 
 [entry_points]
 console_scripts =
-    moteino-api=moteino_sensors.api.api:main
-    moteino-executor=moteino_sensors.executor.executor:main
-    moteino-dbsm=moteino_sensors.dbsm.dbsm:main
-    moteino-srl=moteino_sensors.srl.srl:main
-    moteino-feeder=moteino_sensors.feeder.feeder:main
+    moteino-api=moteino_sensors.api:main
+    moteino-executor=moteino_sensors.executor:main
+    moteino-dbsm=moteino_sensors.dbsm:main
+    moteino-srl=moteino_sensors.srl:main
+    moteino-feeder=moteino_sensors.feeder:main
 
 [wheel]
 universal = 1
